requirements.txt
fastapi==0.99.1
pydantic==1.10.11
python-dotenv==1.0.0
python_json_logger==2.0.7
sentry-sdk==1.27.1
SQLAlchemy==2.0.18
starlette==0.27.0
minio==7.1.15
python-multipart==0.0.6

uvicorn==0.22.0

httpx==0.24.1
asyncpg~=0.28.0
pytest~=7.2.2
alembic~=1.11.1
